misread 'average' problem

+++ funs.c
@@ -1,13 +1,13 @@
 #include "funs.h"
 
 // I just like adding length. Seems more safe.
-int average(int *arr, int length) {
+float average(int *arr, int length) {
   int i;
   float result = 0.0;
   for (i = 0; i < length; i++) {
     result += *(arr + i);
   }
-  return result / length;
+  return (float) result / length;
 }
 
 // Assumes the user is imputing correct lengths.
