+++ funs.h
@@ -1,6 +1,6 @@
 #ifndef FUNS
 #define FUNS
-int average(int *arr, int length);
+float average(int *arr, int length);
 void copy(int *arr, int *ray, int length);
 int len(char *word);
 #endif /* ifndef FUNS */
